Split health_check.go into smaller parts.
Distinct files for core, http, and tcp.
